requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
get==2019.4.13
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
post==2019.4.13
psycopg2-binary==2.8.2
public==2019.4.13
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
six==1.12.0
Werkzeug==0.15.2
